refactor: type checking and split files
