+++ geos-trame/src/geos_trame/app/ui/viewer/regionViewer.py
@@ -10,10 +10,10 @@ class RegionViewer:
 
     This mesh is represented in GEOS with a Region.
     """
+    input: pv.UnstructuredGrid
+    clip: pv.UnstructuredGrid
 
     def __init__( self ) -> None:
-        self.input: pv.UnstructuredGrid
-        self.clip: pv.UnstructuredGrid
         self.reset()
 
     def __call__( self, normal: tuple[ float ], origin: tuple[ float ] ) -> None:
